Combine typed data frames row-wise

This function combines multiple typed data frames row-wise, ensuring type consistency and applying row validation rules. It extends the base rbind function by adding type checks and row validation based on the specified rules for typed data frames.

Usage

# S3 method for class 'typed_frame'
rbind(..., deparse.level = 1)

Arguments

...

Typed data frames to combine.

deparse.level

See rbind.

Value

The combined typed data frame.

Details

This version of rbind for typed_frame performs extra type checking and row validation to ensure consistency and adherence to specified rules. Refer to the base rbind documentation for additional details on combining data frames: rbind.